feat(monitoring): configurable stats bar position (header or bottom)
- new navbarPosition setting in hostNavbarDisplay (top|bottom, default
  bottom), persisted via monitoring/display-settings with validation
- SystemStatsBar component shared by both placements (variant prop);
  Layout renders it in the header row or as the fixed bottom bar
- monitoring tab: position selector next to the existing display toggles,
  i18n wording made position-neutral (en/fr)
