+++ frontend/src/layouts/Layout.vue
@@ -34,7 +34,7 @@
         </div>
 
         <!-- Desktop header -->
-        <header v-if="! $root.isMobile" class="desktop-header py-3 mb-3 border-bottom">
+        <header v-if="! $root.isMobile" class="desktop-header py-3 mb-3 border-bottom" :class="{ 'has-stats': showHeaderStats }">
             <div class="desktop-brand d-flex align-items-center">
                 <router-link to="/" class="d-flex align-items-center text-dark text-decoration-none">
                     <object class="bi me-2 ms-4" width="40" height="40" data="/icon.svg" />
@@ -134,6 +134,14 @@
                     </div>
                 </li>
             </ul>
+
+            <SystemStatsBar
+                v-if="showHeaderStats"
+                variant="header"
+                :system-stats="systemStats"
+                :kula-url="kulaUrl"
+                :dozzle-url="dozzleUrl"
+            />
         </header>
 
         <!-- Mobile header -->
@@ -212,51 +220,19 @@
         </main>
 
         <!-- System status bar (VS Code-style slim bottom bar, desktop/tablet only) -->
-        <footer v-if="$root.loggedIn && systemStats" class="status-bar">
-            <span class="status-item" :class="statClass(systemStats.cpu)" :title="cpuStatTooltip()">
-                <font-awesome-icon icon="microchip" />CPU {{ systemStats.cpu }}%
-                <span class="status-meter" :aria-label="diskUsageBarLabel(systemStats.cpu)">
-                    <span class="status-meter-fill" :style="{ width: Math.min(100, systemStats.cpu) + '%' }"></span>
-                </span>
-            </span>
-            <span class="status-item" :class="statClass(systemStats.ram.percent)" :title="ramStatTooltip()">
-                <font-awesome-icon icon="memory" />RAM {{ systemStats.ram.percent }}%
-                <span class="status-meter" :aria-label="diskUsageBarLabel(systemStats.ram.percent)">
-                    <span class="status-meter-fill" :style="{ width: Math.min(100, systemStats.ram.percent) + '%' }"></span>
-                </span>
-            </span>
-            <span v-if="fullestDisk" class="status-item" :class="statClass(fullestDisk.percent)" :title="disksTooltip">
-                <font-awesome-icon icon="floppy-disk" />{{ fullestDisk.mount }} {{ fullestDisk.percent }}%
-                <span class="status-meter" :aria-label="diskUsageBarLabel(fullestDisk.percent)">
-                    <span class="status-meter-fill" :style="{ width: Math.min(100, fullestDisk.percent) + '%' }"></span>
-                </span>
-                <span v-if="systemStats.diskDisplayMode === 'bar'">{{ formatDiskTotal(fullestDisk.total) }}</span>
-                <span v-if="diskList.length > 1">+{{ diskList.length - 1 }}</span>
-            </span>
-            <span v-if="systemStats.hostNavbarDisplay?.uptime" class="status-item stat-neutral">
-                <font-awesome-icon icon="clock" />{{ formatUptime(systemStats.host?.uptimeSeconds) }}
-            </span>
-            <span v-if="systemStats.hostNavbarDisplay?.cpuTemperatures && systemStats.host?.temperatures?.cpu?.length" class="status-item stat-neutral">
-                <font-awesome-icon icon="temperature-half" />{{ tempSummary(systemStats.host.temperatures.cpu) }}
-            </span>
-            <span v-if="systemStats.hostNavbarDisplay?.diskTemperatures && systemStats.host?.temperatures?.disks?.length" class="status-item stat-neutral">
-                <font-awesome-icon icon="hard-drive" />{{ tempSummary(systemStats.host.temperatures.disks) }}
-            </span>
-
-            <span class="status-bar-spacer"></span>
-
-            <a v-if="kulaUrl" :href="kulaUrl" target="_blank" class="status-item status-link">
-                <font-awesome-icon icon="chart-bar" />Kula
-            </a>
-            <a v-if="dozzleUrl" :href="dozzleUrl" target="_blank" class="status-item status-link">
-                <font-awesome-icon icon="terminal" />Dozzle
-            </a>
-        </footer>
+        <SystemStatsBar
+            v-if="$root.loggedIn && systemStats && statsPosition === 'bottom'"
+            variant="bottom"
+            :system-stats="systemStats"
+            :kula-url="kulaUrl"
+            :dozzle-url="dozzleUrl"
+        />
     </div>
 </template>
 
 <script>
 import Login from "../components/Login.vue";
+import SystemStatsBar from "../components/SystemStatsBar.vue";
 import { compareVersions } from "compare-versions";
 import { ALL_ENDPOINTS } from "../../../common/util-common";
 import { setLowPower, POLL, makePoller } from "../composables/useLowPower";
@@ -266,6 +242,7 @@ export default {
 
     components: {
         Login,
+        SystemStatsBar,
     },
 
     data() {
@@ -310,23 +287,13 @@ export default {
             return `docker pull ghcr.io/${repo}:latest && docker compose up -d`;
         },
 
-        // Barre de stats : on n'affiche que le disque le plus plein,
-        // le détail de tous les disques est dans l'infobulle.
-        diskList() {
-            if (!this.systemStats) {
-                return [];
-            }
-            return this.systemStats.disks ?? (this.systemStats.disk ? [ this.systemStats.disk ] : []);
+        // Where the system stats render (see MonitoringTab display settings)
+        statsPosition() {
+            return this.systemStats?.hostNavbarDisplay?.navbarPosition === "top" ? "top" : "bottom";
         },
 
-        fullestDisk() {
-            return this.diskList.reduce((max, d) => (!max || d.percent > max.percent ? d : max), null);
-        },
-
-        disksTooltip() {
-            return this.diskList
-                .map((d) => `${d.mount} ${d.percent}% (${this.formatDiskTotal(d.total)})`)
-                .join("\n");
+        showHeaderStats() {
+            return this.$root.loggedIn && !!this.systemStats && this.statsPosition === "top";
         },
 
     },
@@ -433,84 +400,6 @@ export default {
             }
         },
 
-        statClass(percent) {
-            if (percent >= 85) return "stat-danger";
-            if (percent >= 70) return "stat-warning";
-            return "stat-ok";
-        },
-
-        formatBytes(bytes) {
-            if (bytes === 0) return "0 B";
-            const gb = bytes / (1024 ** 3);
-            if (gb >= 1) return gb.toFixed(1) + " GB";
-            const mb = bytes / (1024 ** 2);
-            return mb.toFixed(0) + " MB";
-        },
-
-        diskUsageBarLabel(percent) {
-            const clamped = Math.max(0, Math.min(100, Number(percent) || 0));
-            const filled = Math.round(clamped / 10);
-            return `[${"⣿".repeat(filled)}${" ".repeat(10 - filled)}]`;
-        },
-
-        formatDiskTotal(bytes) {
-            if (!bytes) return "0B";
-            const units = ["B", "Kio", "Mio", "Gio", "Tio", "Pio"];
-            let value = bytes;
-            let unitIndex = 0;
-            while (value >= 1024 && unitIndex < units.length - 1) {
-                value /= 1024;
-                unitIndex += 1;
-            }
-            const precision = value >= 10 || unitIndex === 0 ? 0 : 1;
-            return `${value.toFixed(precision)}${units[unitIndex]}`;
-        },
-
-        coreSummary(values) {
-            return values.map((value, index) => `C${index + 1} ${value}%`).join(" ");
-        },
-
-        cpuStatTooltip() {
-            const details = [];
-            if (this.systemStats?.hostNavbarDisplay?.cpuModel && this.systemStats?.host?.cpuModel) {
-                details.push(this.systemStats.host.cpuModel);
-            }
-            if (this.systemStats?.hostNavbarDisplay?.perCoreCpu && this.systemStats?.host?.perCoreCpu?.length) {
-                details.push(this.coreSummary(this.systemStats.host.perCoreCpu));
-            }
-            return details.join("\n");
-        },
-
-        ramStatTooltip() {
-            const ram = this.systemStats?.ram;
-            if (!ram) {
-                return "";
-            }
-            return `${this.formatBytes(ram.used)} / ${this.formatBytes(ram.total)}`;
-        },
-
-        formatUptime(seconds) {
-            const total = Math.max(0, Number(seconds) || 0);
-            const days = Math.floor(total / 86400);
-            const hours = Math.floor((total % 86400) / 3600);
-            if (days > 0) return `${this.$t("timeUnit.day", [ days ])} ${this.$t("timeUnit.hour", [ hours ])}`;
-            return this.$t("timeUnit.hour", [ hours ]);
-        },
-
-        tempSummary(values) {
-            if (!Array.isArray(values) || values.length === 0) {
-                return "";
-            }
-            if (values.length === 1) {
-                return `${values[0].celsius}°C`;
-            }
-            const numbers = values.map(v => Number(v.celsius)).filter(Number.isFinite);
-            if (numbers.length === 0) {
-                return "";
-            }
-            return `${Math.min(...numbers)}-${Math.max(...numbers)}°C`;
-        },
-
         async fetchKulaStatus() {
             try {
                 const token = localStorage.getItem("token") ?? sessionStorage.getItem("token") ?? "";
@@ -584,6 +473,12 @@ export default {
     align-items: center;
     gap: 0.75rem 1.25rem;
     padding-inline: 1.5rem;
+
+    &.has-stats {
+        grid-template-areas:
+            "brand updates navigation"
+            "stats stats stats";
+    }
 }
 
 .desktop-brand {
@@ -807,72 +702,6 @@ main {
         color: var(--warning);
         border-color: color-mix(in srgb, var(--warning) 70%, transparent);
     }
-}
-
-// Slim bottom status bar (VS Code-style). Hidden on mobile.
-.status-bar {
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    z-index: 900; // below mobile header (1000) and offcanvas backdrop (1040)
-    display: flex;
-    align-items: center;
-    gap: var(--space-4);
-    height: 28px;
-    padding: 0 var(--space-4);
-    background: var(--bg-surface);
-    border-top: 1px solid var(--border-color);
-    font-size: var(--fs-xs);
-    font-weight: 500;
-    white-space: nowrap;
-    overflow-x: auto;
-
-    @media (max-width: $bp-mobile) {
-        display: none;
-    }
-}
-
-.status-bar-spacer {
-    flex: 1;
-}
-
-.status-item {
-    display: inline-flex;
-    align-items: center;
-    gap: var(--space-1);
-    transition: color 0.3s;
-
-    &.stat-ok      { color: var(--success); }
-    &.stat-warning  { color: var(--warning); }
-    &.stat-danger   { color: var(--danger); }
-    &.stat-neutral  { color: var(--text-color); }
-}
-
-.status-link {
-    color: var(--primary-strong);
-    text-decoration: none;
-
-    &:hover {
-        color: var(--primary-hover);
-    }
-}
-
-.status-meter {
-    display: inline-block;
-    width: 3rem;
-    height: 4px;
-    border-radius: var(--radius-pill);
-    background: var(--bg-raised);
-    overflow: hidden;
-}
-
-.status-meter-fill {
-    display: block;
-    height: 100%;
-    border-radius: var(--radius-pill);
-    background: currentColor;
-    transition: width 0.3s ease;
 }
 
 .nav {
